Updated calendar controller
Changed around trhe way that the data was stored and presented in the
calendar controller and its view to be hashes instead of lists.
Refactored the event and section method names, and made the variables
just in general more readable. End result is that the index now displays
the sections and events (plus time frames) for the user chosen.