Refactored Seeds File
Made everything look a lot more readable. Also changed it so that the
code is more easily changeable and made it so that the data is all
consistent. So, now when someone looks up for example the entries in the
gradebook, currently there are only 24 users displayed but all of the
section and event data is consistent in every table that the data is
linked to.