Updated Seeds File
Fixed a bug in the seeds file where we were using the reserved keyword
'type' for a column name. Normally this would be used to automatically
infer inheritance of data. Basically to fix that bug we just changed the
column name to something else that was equally descriptive.

Moved the section list getting method to the user model.

Overrode the section model's as_json method, generating the hash of data
of the section id to a list of the events' data, containing the event
name, the title, and the end time each.