Adding DragonDrop Changes to DragonDrop
A lot of changes have been made to the dragon_drop controller and they
haven't been in the dragon_drop branch, so with this commit they now
are. In addition, got the get_courses method to output json data of the
course names. This can be changed later if needed. Still working on the
combo box.