Changes to Seed and Section
Minor changes to DragonDrop. Added a lot of helper methods and did
refactoring on methods in the Section and User models. In addition, also
did a structural change to Section so that it now has the data of what
days of the week it can meet on. This will be used to select which class
day the event is due on.