Changes to the Seeds File
Added another helper method. In additon, the general changes to the seed
file from before were that I reduced the number of users and allowed the
gradebook to be created with all of the users, all of their sections,
and all of the section's assigned events.

Before that, I also made it so that UserSection was creating more
realistic data; previously it was just assigning random sections to
every user including professors, but now it will assign every section to
professors and four random ones to students.

The last change was to SectionEvent, which, based on the new data from
Section of the meeting days, allows it to randomly select a meeting day
for the event to be due on during any random week of the current
semester.