DragonDrop Filters Sections using Semesters
Added in a user model which checks sections against a particular
semester and particular course and rejects sections which do not fit
that criteria. Also changed the seed so that there are more sections so
that there is less of a chance for there to be no sections (and thus no
events) for a particular semester.