An Obscure Fix
Since we changed the way the semester_id was being handled in section
the way to get the semester for it was out of date, so that had to be
changed. In addition, some of the changes to the seed which would have
allowed for some nice data to be created that were overridden or
something. Added those back to the seed.